+++ src/main/java/com/kt/security/JwtAuthenticationFilter.java
@@ -2,8 +2,8 @@ package com.kt.security;
 
 import java.io.IOException;
 
-import com.kt.common.CustomException;
-import com.kt.common.ErrorCode;
+import com.kt.common.exception.CustomException;
+import com.kt.common.exception.ErrorCode;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.util.StringUtils;
